third_party/WebKit/LayoutTests/platform/win/fast/forms/color/display-none-input-color-chooser-shown-expected.txt

This is a testharness.js-based test.
PASS window.eventSender is required for the test to run 
PASS Tests click on label for color picker with display none should show chooser 
Harness: the test ran to completion.
